requirements.txt
# ── Data ───────────────────────────────────────────────────────
yfinance>=0.2.31
pandas>=2.0
numpy>=1.24

# ── Broker (Alpaca paper trading) ─────────────────────────────
alpaca-py>=0.21

# ── Security / Config ─────────────────────────────────────────
python-dotenv>=1.0

# ── Visualisation & Reporting ─────────────────────────────────
matplotlib>=3.7
tabulate>=0.9

# ── Scheduling (for the daily bot loop) ───────────────────────
schedule>=1.2
